+++ Intent.Modules.TypeORM.Entities/Templates/Repository/RepositoryTemplateRegistration.cs
@@ -3,9 +3,11 @@ using System.Collections.Generic;
 using System.Linq;
 using Intent.Engine;
 using Intent.Metadata.Models;
+using Intent.Metadata.RDBMS.Api;
 using Intent.Modelers.Domain.Api;
 using Intent.Modules.Common;
 using Intent.Modules.Common.Registrations;
+using Intent.Modules.TypeORM.Entities.Api;
 using Intent.RoslynWeaver.Attributes;
 using Intent.Templates;
 
@@ -34,7 +36,9 @@ namespace Intent.Modules.TypeORM.Entities.Templates.Repository
         [IntentManaged(Mode.Merge, Body = Mode.Ignore, Signature = Mode.Fully)]
         public override IEnumerable<ClassModel> GetModels(IApplication application)
         {
-            return _metadataManager.Domain(application).GetClassModels().Where(x => x.IsAggregateRoot());
+            return _metadataManager.Domain(application).GetClassModels()
+                .Where(x => (x.IsAggregateRoot() && (!x.IsAbstract || x.HasTable())) || x.HasRepository())
+                .ToArray();
         }
     }
 }